--- src/components/login/index.js
@@ -1,5 +1,0 @@
-import * as selectors from './selectors';
-
-export { selectors };
-
-export { default } from './reducers';
