+++ src/pages/login/components/index.js
@@ -1,0 +1,3 @@
+export { default as LoginDonor } from './LoginDonor';
+export { default as LoginLanding } from './LoginLanding';
+export { default as LoginNpo } from './LoginNpo';
